requirements.txt
httpx==0.27.0
multiprocess==0.70.16
openai==1.40.1
pandas==2.2.2
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
uuid_utils==0.9.0
